refactor(testcafe): extract type interfaces into src/types.ts
